chore(v2): gate 4 cleanup (clippy, leak deferral note)
Closes out Gate 4 after commits V and W landed the performance work.
Commit X (epoch reclamation for dep lists) was dropped from Gate 4;
see devlog and the expanded comment on replace_deps_leaking_old_overflow
for the rationale.

Changes:

- Fix v2 clippy warnings under Rust 1.91's stricter lints:
  - Remove duplicated `#![cfg(test)]` inner attributes in
    runtime_concurrent_test.rs, runtime_proptest.rs, and
    runtime_vs_v1_bench.rs (the outer `#[cfg(test)]` in mod.rs
    already gates these modules).
  - Remove unused `UnsafeCell` import in arena.rs (leftover from
    commit U's mutex-guarded GenericSlot rewrite).
  - Remove unnecessary `as u32` cast in nodes_store.rs's drop-leak
    test.
  - Simplify `0 + 1 + 2` to `1 + 2` in runtime.rs's
    conditional-deps test (identity_op lint).

- Expand the comment on `NodeData::replace_deps_leaking_old_overflow`
  to explicitly document that the leak is kept as the permanent
  post-Gate-4 state, with the crossbeam-epoch 0.9 miri finding and
  the deferred-work pointer to evaluate seize / haphazard / custom
  strict-provenance reclamation in a later dedicated chunk.

Gate 4 final numbers (5-run median, v2/v1 ratio, clean environment):

| Benchmark                                | U ratio | Y ratio | Delta |
|------------------------------------------|---------|---------|-------|
| hot_clean_read_input_u64                 | 2.16x   | 1.59x   | -26%  |
| hot_clean_read_query_u64                 | 2.15x   | 1.72x   | -20%  |
| set_input_no_dependents                  | 2.01x   | 1.69x   | -16%  |
| cold_build_chain_1000 per node           | 1.18x   | 1.22x   | noise |
| propagate_chain_1000                     | 5.90x   | 1.92x   | -67%  |
| propagate_chain_1000_stable_intermediate | (new)   | 3.57x   | (new) |

Headline: propagate dropped from 5.90x v1 to 1.92x v1, the largest
single improvement in the chunk. The stable_intermediate benchmark is
new in W; v2 has red-green but pays more per-level than v1 because
of atomic-op and RwLock-guard overhead. Closing that gap is queued
as follow-up post-Gate-4 work.

Verification:
- cargo test -p incr-core: all 224 tests pass
- cargo clippy -p incr-core --tests: v2 module clean (remaining v1
  warnings are pre-existing and out of scope for Gate 4)
- miri targeted: concurrent_get_and_set_on_vec_input_is_race_free
  (Vec C1 race, 125s). Clean.

Full chunk narrative in devlogs/2026-04-06-gate-4-performance-parity.md.

+++ crates/incr-core/src/v2/nodes_store.rs
@@ -261,7 +261,7 @@ mod tests {
         let store = SegmentedNodes::new();
         for i in 0..20 {
             let slot = store.push(NodeData::new_query(0, i));
-            let deps: Vec<NodeId> = (0..(i as u32 + 5)).map(NodeId).collect();
+            let deps: Vec<NodeId> = (0..(i + 5)).map(NodeId).collect();
             store.get(slot).publish_initial_deps(&deps);
         }
         drop(store);
